WIP: making SDXL run on TPU v2
I manually place weights and inputs on specific devices, instead of using
`pjit`, like follows:
- Text encoders -> device 0
- Base UNet -> devices [1, 2]
- Refiner UNet -> devices [3, 4]
- VAE -> device 7

This suffices to generate 2 images. I don’t use devices 5 and 6 for the
unets because the vae OOMs decoding 3 images in a single device. I could
use devices [6, 7] for the vae and use a batch size of 2 in the UNets to
generate 4 images.

The process works, but the refiner outputs seem wrong – this would be an
unrelated bug to be investigated.

This will be slow because of the old hardware and all the memory
transfers. I'll measure speed and increase batch size after the refiner
bug is solved.